LayoutTests/http/tests/security/contentSecurityPolicy/upgrade-insecure-requests/basic-upgrade-cors.https-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/contentSecurityPolicy/upgrade-insecure-requests/resources/basic-upgrade-cors.https.html requested insecure content from http://127.0.0.1:8443/security/resources/abe.png. This content was automatically upgraded and should be served over HTTPS.

CONSOLE MESSAGE: PASS: Successfully retrieved image data.
This test opens a window that loads an insecure image. We should upgrade this request and thereby avoid triggering a mixed content callback.



PASS Verify that images have correct cross-origin behavior.
